+++ bot_guide.txt
